+++ tests/optimization/test_tiktak.py
@@ -3,7 +3,6 @@ from itertools import product
 import numpy as np
 import pandas as pd
 import pytest
-from estimagic.optimization.optimize import process_multistart_sample
 from estimagic.optimization.tiktak import _linear_weights
 from estimagic.optimization.tiktak import _tiktak_weights
 from estimagic.optimization.tiktak import draw_exploration_sample
@@ -25,28 +24,6 @@ def params():
 @pytest.fixture
 def constraints():
     return [{"type": "fixed", "loc": "c", "value": 2}]
-
-
-samples = [
-    (
-        pd.DataFrame(np.ones((2, 3)), columns=["a", "b", "c"]),
-        pd.Series([1, 2, 3], index=["a", "b", "c"], name="value").to_frame(),
-        lambda x: x["value"].to_numpy(),
-    ),
-    (
-        np.ones((2, 3)),
-        np.array([1, 2, 3]),
-        lambda x: x,
-    ),
-]
-
-
-@pytest.mark.parametrize("sample, x, to_internal", samples)
-def test_process_multistart_sample(sample, x, to_internal):
-
-    calculated = process_multistart_sample(sample, x, to_internal)
-    expeceted = np.ones((2, 3))
-    aaae(calculated, expeceted)
 
 
 dim = 2
